refactor(bus): split from_config into its own module
Move SystemBus::from_config (~460 lines, the bus builder) out of the
3509-line bus/mod.rs into bus/from_config.rs as a separate impl block. The
private helpers it calls (canonical_peripheral_type, parse_profile_or_default,
resolve_peripheral_path, resolve_pin_*, yaml_*, chip_has_bit_band) stay in
mod.rs — a child module sees its ancestors' private items, so no visibility
changes were needed.

bus/mod.rs 3509 -> 3046. Pure move — register_coverage (all 15 chips) +
workspace lib green, clippy clean.
